5644 client - Move the MCP components shared with embedded into the platform layer
