+++ client/src/ee/pages/embedded/mcp-servers/components/mcp-component-list/McpComponentToolListItem.tsx
@@ -1,7 +1,7 @@
 import Button from '@/components/Button/Button';
 import DeleteAlertDialog from '@/components/DeleteAlertDialog';
 import {Popover, PopoverAnchor} from '@/components/ui/popover';
-import McpComponentToolPropertiesPopover from '@/pages/automation/mcp-servers/components/mcp-component-list/McpComponentToolPropertiesPopover';
+import McpComponentToolPropertiesPopover from '@/pages/platform/mcp-servers/components/McpComponentToolPropertiesPopover';
 import {useCloseActivePopoverOnUnmount, useMcpActivePopover} from '@/shared/contexts/McpActivePopoverContext';
 import {McpTool} from '@/shared/middleware/graphql';
 import {BoltIcon, Trash2Icon} from 'lucide-react';
