+++ client/src/pages/platform/mcp-servers/components/McpComponentSelectionGrid.tsx
