+++ client/src/pages/platform/mcp-servers/components/McpToolListSkeleton.tsx
