+++ client/src/pages/platform/mcp-servers/hooks/useMcpComponentToolPropertiesPopover.ts
